Priority: -Timer for lagouts, timer for subs currently 60 and 30 secs respectivily Done (60 secs from the of lagout) 30 sec for subs and switches from time of using !sub + !switch command -Bug fix to make bot force flag switches - Needs to be done -Increased Stats stored to database -Kills and subs in database for castro - Needs to be done
Gold Plating:
-Dual MVP, offensive + defensive ship
-*warn when ppl talk in blueout Done
-TWL egible needs to be checked //i believe low on the list, still runs with TWD roster
-Increased stats to load for reloading game //since we're not reloading stuff

bug fixes:

Priority:
- Assign caps after line-ups are in (use the !t#-setcaptain options)
- Remove the option to add a player after !startgame Done
Regarding Extension: The extension time is implemented as follows. I think this is a good idea and can be changed if opposed.
The team has to have minimum by the 10 mins required. NO IFS OR BUTS. However, they can add players futher to the 2 mins after game starts to make up the maximum.

Otherwise, what is the point of the extension you should just make the total time for lineups itself 12 mins. There has to be a difference in the condition between extension time and regular time.

Gold Plating
- blueout still doesnt work, twl is a spec chat only blueout (this is fucking tricky and we may just want to remove it completely. There are two commands to get it working, you have to *lockpublic then *lockspec)
- Place players who are subbed out on their team freq Done
- Add in a shipchange option (onlyone shipchange), but do not enable at this tame


git-svn-id: http://www.twcore.org/svn/trunk/twcore@111 9f8b3893-2a26-0410-9d00-a71e40fd2f77