Priority: -Timer for lagouts, timer for subs currently 60 and 30 secs respectivily Done (60 secs from the of lagout) 30 sec and 15 sec for subs and switches from time of using !sub + !switch command -Bug fix to make bot force flag switches - Semi Done, needs to be tested to see results, I dont think this can be solved! We can no way know what other clients are thinking the fix will only make sure ppl know when matchbot see flagClaimed. But if someone lags onto the flag and matchbot doesnt see it and the turf_packet conincides with matchbot then it cannot rectify the other clients vision.
Since this is already there no matter what type of flag game you play, you have to either talk to priitk to change client prediction or something.
-Increased Stats stored to database Done
-Kills and subs in database for castro - Needs to be done

Gold Plating:
-Dual MVP, offensive + defensive ship
-*warn when ppl talk in blueout Done
-TWL egible needs to be checked //i believe low on the list, still runs with TWD roster
-Increased stats to load for reloading game //since we're not reloading stuff

bug fixes:

Priority:
- Assign caps after line-ups are in (use the !t#-setcaptain options)Done
!setcaptain is the command and it differentiates the player by his squad which it looks up from the dbase like !add player
- Remove the option to add a player after !startgame Done
Regarding Extension: The extension time is implemented as follows. I think this is a good idea and can be changed if opposed.
The team has to have minimum by the 10 mins required. NO IFS OR BUTS. However, they can add players futher to the 2 mins after game starts to make up the maximum.

Otherwise, what is the point of the extension you should just make the total time for lineups itself 12 mins. There has to be a difference in the condition between extension time and regular time.

Gold Plating
- blueout still doesnt work, twl is a spec chat only blueout (this is fucking tricky and we may just want to remove it completely. There are two commands to get it working, you have to *lockpublic then *lockspec)
Done
This was already done, there was a bug in how it was implemented. Hopefully I fixed it
- Place players who are subbed out on their team freq Done


git-svn-id: http://www.twcore.org/svn/trunk@116 9f8b3893-2a26-0410-9d00-a71e40fd2f77