Reverting changeset [2473] of dugwyler. I'm sorry, I should've committed this change earlier (I forgot to remove EventRequest.java file earlier).
The EventRequest class has been copied into the eventbot.java file on purpose so it's certain the class is loaded when the bot is spawned. 
I've found out with the pubstats that classes/files outside the bot file aren't loaded when the bot is spawned (probably because the bot class is loaded dynamically).

git-svn-id: http://www.twcore.org/svn/trunk/twcore@2476 9f8b3893-2a26-0410-9d00-a71e40fd2f77