Warp update. Fixed some very poorly-written code. 4 checks to remove players when in Levi (not 1, 4! wtf!). Buggy warping algorithm that removed anyone on private frequency from the warp list, regardless of if they'd want to still be warped in when on a public later.
It worked fine the first time I wrote it... ofc not the *very* first time (hardly!). And we all make errors. But I generally like to finish what I start. Please be so kind as to do the same next time ... after you decide to change perfectly good code, and replace it with bug-riddled bullshit -- or at least have the decency to revert!

-qan

git-svn-id: http://www.twcore.org/svn/trunk/twcore@7410 9f8b3893-2a26-0410-9d00-a71e40fd2f77